Removed anonymous function wrapping retFn default.
Because the body of the anonymous function only called another function which was passed the same argument as the wrapper, the wrapper was actually redundant.  Instead you can just pass a reference to the function you want to call, and it will be executed with the proper arguments at runtime.